+++ lib/src/lib.dart
@@ -1,56 +1,36 @@
 import 'dart:ffi';
-import 'dart:io';
 
 import 'package:ffi/ffi.dart';
 
-import 'libserialport_bindings.g.dart';
-
-const String _libName = 'libserialport_plus';
-
-/// The dynamic library in which the symbols for [FlutterSerialBindings] can be found.
-final DynamicLibrary _dylib = () {
-  if (Platform.isMacOS || Platform.isIOS) {
-    return DynamicLibrary.open('$_libName.framework/$_libName');
-  }
-  if (Platform.isAndroid || Platform.isLinux) {
-    return DynamicLibrary.open('lib$_libName.so');
-  }
-  if (Platform.isWindows) {
-    return DynamicLibrary.open('$_libName.dll');
-  }
-  throw UnsupportedError('Unknown platform: ${Platform.operatingSystem}');
-}();
-
-/// The bindings to the native functions in [_dylib].
-final LibSerialPortBindings lib = LibSerialPortBindings(_dylib);
+import 'libserialport_bindings.g.dart' as sp;
 
 int assertReturn(int value) {
   if (value >= 0) return value;
-  if (value == sp_return.FAIL) {
+  if (value == sp.Return.ERR_FAIL) {
     final ex = getLastError();
     if (ex != null) throw ex;
   }
   String message = switch (value) {
-    sp_return.ARG => 'Argument error',
-    sp_return.FAIL => 'Fail',
-    sp_return.MEM => 'Memory error',
-    sp_return.SUPP => 'Unsupported',
+    sp.Return.ERR_ARG => 'Argument error',
+    sp.Return.ERR_FAIL => 'Fail',
+    sp.Return.ERR_MEM => 'Memory error',
+    sp.Return.ERR_SUPP => 'Unsupported',
     _ => 'Unknown error',
   };
   throw SerialPortException(value, message);
 }
 
 SerialPortException? getLastError() {
-  final code = lib.last_error_code();
+  final code = sp.lastErrorCode();
   if (code == 0) return null;
-  final ptr = lib.last_error_message();
+  final ptr = sp.lastErrorMessage();
   try {
     return SerialPortException(
       code,
       ptr != nullptr ? ptr.cast<Utf8>().toDartString() : '',
     );
   } finally {
-    lib.free_error_message(ptr);
+    sp.freeErrorMessage(ptr);
   }
 }