+++ src/include/config.h
@@ -1,12 +1,14 @@
-/* config.h.  Generated from config.h.in by configure.  */
-/* config.h.in.  Generated from configure.ac by autoheader.  */
+/* config.h values used when building libserialport through Dart native assets. */
 
 /* clock_gettime is available. */
 #define HAVE_CLOCK_GETTIME 1
 
-/* Define to 1 if you have the declaration of `BOTHER', and to 0 if you don't.
- */
+/* Define to 1 if you have the declaration of `BOTHER', and to 0 if you don't. */
+#if defined(__APPLE__)
+#define HAVE_DECL_BOTHER 0
+#else
 #define HAVE_DECL_BOTHER 1
+#endif
 
 /* Define to 1 if you have the <dlfcn.h> header file. */
 #define HAVE_DLFCN_H 1
@@ -26,6 +28,9 @@
 /* Define to 1 if you have the <stdint.h> header file. */
 #define HAVE_STDINT_H 1
 
+/* Define to 1 if you have the <stdio.h> header file. */
+#define HAVE_STDIO_H 1
+
 /* Define to 1 if you have the <stdlib.h> header file. */
 #define HAVE_STDLIB_H 1
 
@@ -36,25 +41,16 @@
 #define HAVE_STRING_H 1
 
 /* Define to 1 if the system has the type `struct serial_struct'. */
+#if defined(__linux__) && !defined(__ANDROID__)
 #define HAVE_STRUCT_SERIAL_STRUCT 1
+#endif
 
 /* Define to 1 if the system has the type `struct termios2'. */
+#if defined(__linux__)
 #define HAVE_STRUCT_TERMIOS2 1
-
-/* Define to 1 if `c_ispeed' is a member of `struct termios2'. */
 #define HAVE_STRUCT_TERMIOS2_C_ISPEED 1
-
-/* Define to 1 if `c_ospeed' is a member of `struct termios2'. */
 #define HAVE_STRUCT_TERMIOS2_C_OSPEED 1
-
-/* Define to 1 if `c_ispeed' is a member of `struct termios'. */
-/* #undef HAVE_STRUCT_TERMIOS_C_ISPEED */
-
-/* Define to 1 if `c_ospeed' is a member of `struct termios'. */
-/* #undef HAVE_STRUCT_TERMIOS_C_OSPEED */
-
-/* Define to 1 if the system has the type `struct termiox'. */
-/* #undef HAVE_STRUCT_TERMIOX */
+#endif
 
 /* sys/file.h is available. */
 #define HAVE_SYS_FILE_H 1
@@ -70,12 +66,6 @@
 
 /* Define to the sub-directory where libtool stores uninstalled libraries. */
 #define LT_OBJDIR ".libs/"
-
-/* Enumeration is unsupported. */
-/* #undef NO_ENUMERATION */
-
-/* Port metadata is unavailable. */
-/* #undef NO_PORT_METADATA */
 
 /* Define to the address where bug reports for this package should be sent. */
 #define PACKAGE_BUGREPORT "martin-libserialport@earth.li"
@@ -128,23 +118,16 @@
 /* Define to 1 if you have the ANSI C header files. */
 #define STDC_HEADERS 1
 
-/* Enable large inode numbers on Mac OS X 10.5.  */
+/* Enable large inode numbers on Mac OS X 10.5. */
 #ifndef _DARWIN_USE_64_BIT_INODE
 #define _DARWIN_USE_64_BIT_INODE 1
 #endif
 
-/* Number of bits in a file offset, on hosts where this is settable. */
-/* #undef _FILE_OFFSET_BITS */
-
-/* Define for large files, on AIX-style hosts. */
-/* #undef _LARGE_FILES */
-
-/* Define to `unsigned int' if <sys/types.h> does not define. */
-/* #undef size_t */
-
-#if HAVE_STRUCT_TERMIOS_C_ISPEED && HAVE_STRUCT_TERMIOS_C_OSPEED
+#if defined(HAVE_STRUCT_TERMIOS_C_ISPEED) && \
+    defined(HAVE_STRUCT_TERMIOS_C_OSPEED)
 #define HAVE_TERMIOS_SPEED 1
 #endif
-#if HAVE_STRUCT_TERMIOS2_C_ISPEED && HAVE_STRUCT_TERMIOS2_C_OSPEED
+#if defined(HAVE_STRUCT_TERMIOS2_C_ISPEED) && \
+    defined(HAVE_STRUCT_TERMIOS2_C_OSPEED)
 #define HAVE_TERMIOS2_SPEED 1
 #endif