+++ test/serial_port_reader_test.dart
@@ -1,6 +1,6 @@
 import 'dart:typed_data';
 
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:libserialport_plus/libserialport_plus.dart';
 
 // Set these to the names of your null-modem serial ports.